🧚‍♀️🧚‍♀️🧚‍♀️RELテーブルの削除

+++ src/main/java/otaku/info/dto/PAllDto.java
@@ -3,7 +3,6 @@ package otaku.info.dto;
 import lombok.Getter;
 import lombok.Setter;
 
-import java.math.BigInteger;
 import java.util.Map;
 import java.util.List;
 
@@ -24,5 +23,5 @@ public class PAllDto {
     private List<RegPMDto> regPmList;
 
     // 各チームの未チェックID件数
-    private Map<BigInteger, BigInteger> pNumberMap;
+    private Map<Long, Integer> pNumberMap;
 }
